# apps/backend/static/openapi.json
{
  "openapi": "3.0.2",
  "paths": {
...
    "/v1/vendors": {
      "post": {
        "description": "Add vendor to database",
        "summary": "Create a new vendor",
        "tags": [
          "Vendor"
        ],
        "parameters": [],
        "operationId": "createVendor",
        "requestBody": {
          "description": "Body",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 150
                  },
                  "about": {
                    "type": "string"
                  },
                  "cuisine": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "deliveryAvailable": {
                    "type": "boolean"
                  },
                  "pickupAvailable": {
                    "type": "boolean"
                  },
                  "groupOrderAvailable": {
                    "type": "boolean"
                  },
                  "deliveryFee": {
                    "type": "number",
                    "minimum": 0
                  },
                  "minOrderAmount": {
                    "type": "number",
                    "minimum": 0
                  },
                  "deliveryTimeEstimate": {
                    "type": "string"
                  },
                  "isOpen": {
                    "type": "boolean"
                  },
                  "openingHours": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "string"
                    }
                  },
                  "vendorListingImageName": {
                    "type": "string"
                  },
                  "vendorLogoImageName": {
                    "type": "string"
                  },
                  "vendorType": {
                    "type": "string",
                    "enum": [
                      "restaurant",
                      "bakery",
                      "alcohol",
                      "cafe"
                    ]
                  },
                  "isFeatured": {
                    "type": "boolean"
                  },
                  "cuisineTags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "promoText": {
                    "type": "string"
                  },
                  "vendorNotice": {
                    "type": "string"
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "201",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "about": {
                      "type": "string"
                    },
                    "cuisine": {
                      "type": "string"
                    },
                    "phone": {
                      "type": "string",
                      "nullable": true
                    },
                    "rating": {
                      "type": "number"
                    },
                    "reviewCount": {
                      "type": "number"
                    },
                    "deliveryAvailable": {
                      "type": "boolean"
                    },
                    "pickupAvailable": {
                      "type": "boolean"
                    },
                    "groupOrderAvailable": {
                      "type": "boolean"
                    },
                    "deliveryFee": {
                      "type": "number"
                    },
                    "minOrderAmount": {
                      "type": "number"
                    },
                    "deliveryTimeEstimate": {
                      "type": "string"
                    },
                    "isOpen": {
                      "type": "boolean"
                    },
                    "openingHours": {
                      "type": "string",
                      "nullable": true
                    },
                    "vendorListingImage": {
                      "type": "string",
                      "nullable": true
                    },
                    "vendorLogoImage": {
                      "type": "string",
                      "nullable": true
                    },
                    "vendorType": {
                      "type": "string",
                      "nullable": true
                    },
                    "favoriteCount": {
                      "type": "number"
                    },
                    "isFeatured": {
                      "type": "boolean"
                    },
                    "cuisineTags": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "promoText": {
                      "type": "string"
                    },
                    "vendorNotice": {
                      "type": "string"
                    },
                    "vendorServiceCharge": {
                      "type": "number"
                    },
                    "vat": {
                      "type": "number"
                    },
                    "vendorDiscount": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "id",
                    "createdAt",
                    "updatedAt",
                    "name",
                    "about",
                    "cuisine",
                    "rating",
                    "reviewCount",
                    "deliveryAvailable",
                    "pickupAvailable",
                    "groupOrderAvailable",
                    "deliveryFee",
                    "minOrderAmount",
                    "deliveryTimeEstimate",
                    "isOpen",
                    "favoriteCount",
                    "isFeatured",
                    "promoText",
                    "vendorNotice",
                    "vendorServiceCharge",
                    "vat",
                    "vendorDiscount"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "summary": "Get list of vendors with pagination/filter",
        "tags": [
          "Vendor"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "number",
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "number",
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "created_at",
                "updated_at",
                "rating",
                "name"
              ]
            }
          },
          {
            "name": "order",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string",
              "minLength": 2
            }
          },
          {
            "name": "cuisine",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isOpen",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "isFeatured",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "minRating",
            "in": "query",
            "schema": {
              "type": "number",
              "minimum": 0,
              "maximum": 5
            }
          }
        ],
        "operationId": "getVendors",
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "createdAt": {
                            "type": "string"
                          },
                          "updatedAt": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "about": {
                            "type": "string"
                          },
                          "cuisine": {
                            "type": "string"
                          },
                          "phone": {
                            "type": "string",
                            "nullable": true
                          },
                          "rating": {
                            "type": "number"
                          },
                          "reviewCount": {
                            "type": "number"
                          },
                          "deliveryAvailable": {
                            "type": "boolean"
                          },
                          "pickupAvailable": {
                            "type": "boolean"
                          },
                          "groupOrderAvailable": {
                            "type": "boolean"
                          },
                          "deliveryFee": {
                            "type": "number"
                          },
                          "minOrderAmount": {
                            "type": "number"
                          },
                          "deliveryTimeEstimate": {
                            "type": "string"
                          },
                          "isOpen": {
                            "type": "boolean"
                          },
                          "openingHours": {
                            "type": "string",
                            "nullable": true
                          },
                          "vendorListingImage": {
                            "type": "string",
                            "nullable": true
                          },
                          "vendorLogoImage": {
                            "type": "string",
                            "nullable": true
                          },
                          "vendorType": {
                            "type": "string",
                            "nullable": true
                          },
                          "favoriteCount": {
                            "type": "number"
                          },
                          "isFeatured": {
                            "type": "boolean"
                          },
                          "cuisineTags": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "promoText": {
                            "type": "string"
                          },
                          "vendorNotice": {
                            "type": "string"
                          },
                          "vendorServiceCharge": {
                            "type": "number"
                          },
                          "vat": {
                            "type": "number"
                          },
                          "vendorDiscount": {
                            "type": "number"
                          }
                        },
                        "required": [
                          "id",
                          "createdAt",
                          "updatedAt",
                          "name",
                          "about",
                          "cuisine",
                          "rating",
                          "reviewCount",
                          "deliveryAvailable",
                          "pickupAvailable",
                          "groupOrderAvailable",
                          "deliveryFee",
                          "minOrderAmount",
                          "deliveryTimeEstimate",
                          "isOpen",
                          "favoriteCount",
                          "isFeatured",
                          "promoText",
                          "vendorNotice",
                          "vendorServiceCharge",
                          "vat",
                          "vendorDiscount"
                        ]
                      }
                    },
                    "total": {
                      "type": "number"
                    },
                    "page": {
                      "type": "number"
                    },
                    "limit": {
                      "type": "number"
                    },
                    "totalPages": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "data",
                    "total",
                    "page",
                    "limit",
                    "totalPages"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/vendors/addresses": {
      "post": {
        "description": "Add vendor address to database",
        "summary": "Create a new vendor address",
        "tags": [
          "Vendor"
        ],
        "parameters": [],
        "operationId": "createVendorAddress",
        "requestBody": {
          "description": "Body",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "vendorId": {
                    "type": "string"
                  },
                  "streetAddress": {
                    "type": "string"
                  },
                  "city": {
                    "type": "string"
                  },
                  "state": {
                    "type": "string"
                  },
                  "zipcode": {
                    "type": "string"
                  },
                  "latitude": {
                    "type": "number"
                  },
                  "longitude": {
                    "type": "number"
                  }
                },
                "required": [
                  "vendorId",
                  "streetAddress",
                  "city",
                  "state",
                  "zipcode",
                  "latitude",
                  "longitude"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "201",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "vendorId": {
                      "type": "string"
                    },
                    "streetAddress": {
                      "type": "string"
                    },
                    "city": {
                      "type": "string"
                    },
                    "state": {
                      "type": "string"
                    },
                    "zipcode": {
                      "type": "string"
                    },
                    "latitude": {
                      "type": "number"
                    },
                    "longitude": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "vendorId",
                    "streetAddress",
                    "city",
                    "state",
                    "zipcode",
                    "latitude",
                    "longitude"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "info": {
    "version": "1.0.0",
    "title": "Boilerplate REST API - Documentation",
    "description": "Boilerplate REST API - Documentation"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "Local Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "x-service-token": {
        "type": "apiKey",
        "name": "x-service-token",
        "in": "header"
      }
    }
  }
}